chore: relocate analyse_face_levels into recognition/ (#191)
#191 ("unify axial-step recognition on find_turned_steps, retire analyse_face_levels")
is obsolete: since #244/#237, step_zs is dispatched by part class — turned parts use
find_turned_steps.shoulders (OD-silhouette filtered), prismatic parts use
analyse_face_levels (a box-stepped part has no cylinders, so find_turned_steps cannot
see its steps). They answer different questions for disjoint part classes; the
duplication #191 described is gone, and find_turned_steps genuinely cannot replace
analyse_face_levels.

The one real residual is architectural: analyse_face_levels is a recognition
primitive that lived in analysis.py, not recognition/ (ADR 0007). Move it to
recognition/levels.py alongside find_turned_steps (turned.py), so both axial-step
recognisers live together. analysis.py re-exports it (used internally + by the public
API), so `from draftwright[.analysis] import analyse_face_levels` is unchanged.

Pure relocation — no behaviour/output change; import identity preserved across all
three paths. Removes 4 now-unused OCP imports from analysis.py. CLAUDE.md's
recognition/ description updated (it omitted turned.py too). Full fast suite 587
passed / 1 skipped; ruff(0.15.17)/format/mypy clean. Closes #191.
